+++ Steepfile
@@ -19,4 +19,5 @@ target :lib do
   ignore "lib/herb/bootstrap.rb"
   ignore "lib/herb/dev/server.rb"
   ignore "lib/herb/dev/server_entry.rb"
+  ignore "lib/herb/dev/runner.rb"
 end
