+++ javascript/packages/client/src/types.ts
@@ -18,11 +18,6 @@ export interface ReloadMessage {
   file: string
 }
 
-export interface RefreshMessage {
-  type: "refresh"
-  file: string
-}
-
 export interface ParseError {
   name: string
   message: string
@@ -46,7 +41,7 @@ export interface WelcomeMessage {
   project: string
 }
 
-export type HerbMessage = WelcomeMessage | PatchMessage | ReloadMessage | RefreshMessage | ErrorMessage | FixedMessage
+export type HerbMessage = WelcomeMessage | PatchMessage | ReloadMessage | ErrorMessage | FixedMessage
 
 export type ConnectionState = "connected" | "disconnected" | "gave-up"
 
@@ -67,7 +62,6 @@ export interface HerbClientOptions {
   port?: number
   host?: string
   onPatch?: (message: PatchMessage) => void
-  onRefresh?: (message: RefreshMessage) => void
   onReload?: (message: ReloadMessage) => void
   onError?: (message: ErrorMessage) => void
   onFixed?: (message: FixedMessage) => void
