+++ lib/herb/cli.rb
@@ -456,391 +456,23 @@ class Herb::CLI
   end
 
   def dev_stop
-    require_relative "dev/server"
-
-    entries = Herb::Dev::ServerEntry.all
-
-    if entries.empty?
-      puts "No herb dev servers running."
-      exit(0)
-    end
-
-    entries.each do |entry|
-      entry.stop!
-      puts "Stopped herb dev server for #{entry.project_name} (PID: #{entry.pid}, port: #{entry.port})"
-    end
-
-    exit(0)
+    require_relative "dev/runner"
+    Herb::Dev::Runner.new.stop
   end
 
   def dev_restart
-    require_relative "dev/server"
-
-    Herb::Dev::ServerEntry.all.each(&:stop!)
-    sleep 0.5
-    run_dev_server
+    require_relative "dev/runner"
+    Herb::Dev::Runner.new(path: @file || ".").restart
   end
 
   def dev_status
-    require_relative "dev/server"
-
-    entries = Herb::Dev::ServerEntry.all
-
-    if entries.empty?
-      puts "No herb dev servers running."
-    else
-      entries.each do |entry|
-        puts "#{entry.project_name} — PID: #{entry.pid}, port: #{entry.port}, started: #{entry.started_at}"
-      end
-    end
-
-    exit(0)
+    require_relative "dev/runner"
+    Herb::Dev::Runner.new.status
   end
 
   def run_dev_server
-    begin
-      require "cruise"
-    rescue LoadError
-      puts "The 'cruise' gem is required for the dev server."
-      puts
-      puts "Install it:"
-      puts "  gem install cruise"
-      puts ""
-      puts "or add to your Gemfile"
-      puts "  bundle add cruise"
-      exit(1)
-    end
-
-    require_relative "dev/server"
-
-    path = @file || "."
-
-    unless File.directory?(path)
-      puts "Not a directory: '#{path}'."
-      exit(1)
-    end
-
-    config = Herb::Configuration.load(path)
-    expanded_path = File.realpath(File.expand_path(config.project_root || path))
-
-    existing = Herb::Dev::ServerEntry.find_by_project(expanded_path)
-
-    if existing
-      puts "Herb dev server is already running for this project (PID: #{existing.pid}, port: #{existing.port})."
-      puts
-      puts "  herb dev stop       Stop the running server"
-      puts "  herb dev restart    Restart the server"
-      exit(1)
-    end
-
-    port = Herb::Dev::Server::DEFAULT_PORT
-    port_owner = Herb::Dev::ServerEntry.find_by_port(port)
-
-    if port_owner
-      port = Herb::Dev::Server.find_available_port(port + 1)
-      abort "No available ports found" unless port
-    end
-
-    print CLEAR_SCREEN
-    print HIDE_CURSOR
-    puts
-    puts fg_bg(" \u{1F33F} Herb Dev Server ", 255, 28)
-    puts
-    puts "  #{fg("\u26A0\uFE0F Experimental:", 214)} #{fg("The dev server is experimental and may not work correctly in all cases.", 241)}"
-    puts
-
-    include_patterns = config.file_include_patterns
-    exclude_patterns = config.file_exclude_patterns
-
-    puts "  #{fg("Herb:".ljust(11), 245)}#{fg(Herb::VERSION, 250)}"
-    puts "  #{fg("Project:".ljust(11), 245)}#{fg(expanded_path, 250)}"
-    puts "  #{fg("PID:".ljust(11), 245)}#{fg(Process.pid, 250)} #{fg("(#{File.join(Herb::Dev::ServerEntry::SERVERS_DIR, "#{Process.pid}.json")})", 241)}"
-
-    if config.config_path
-      relative_config = config.config_path.to_s.delete_prefix("#{expanded_path}/")
-      puts "  #{fg("Config:".ljust(11), 245)}#{fg(relative_config, 250)}"
-    else
-      puts "  #{fg("Config:".ljust(11), 245)}#{fg("(defaults)", 241)}"
-    end
-
-    puts "  #{fg("Indexing files...", 241)}"
-
-    file_states = {}
-    initial_files = config.find_files(path)
-
-    initial_files.each do |file_path|
-      file_states[file_path] = File.read(file_path)
-    rescue StandardError
-      # skip files that can't be read
-    end
-
-    print "\e[1A\e[2K"
-    puts "  #{fg("Files:".ljust(11), 245)}#{fg("#{file_states.size} templates indexed", 250)}"
-
-    websocket = Herb::Dev::Server.new(port: port, project_path: expanded_path)
-    websocket.start
-
-    puts "  #{fg("WebSocket:".ljust(11), 245)}#{fg("ws://localhost:#{websocket.port}", 250)}"
-    puts
-    puts "  #{fg("Ready!", 42)} #{fg("Watching for changes...", 241)}"
-    puts
-
-    first_change = true
-    errored_files = Set.new
-
-    Thread.new do
-      $stdin.gets(nil)
-      Thread.main.raise(Interrupt)
-    end
-
-    Cruise.watch(expanded_path, only: ["created", "modified", "removed"]) do |event|
-      file_path = event.path
-      relative_path = file_path.delete_prefix("#{expanded_path}/")
-
-      next if config.path_excluded?(relative_path, exclude_patterns)
-      next unless config.path_included?(relative_path, include_patterns)
-
-      if first_change
-        print "\e[2A\e[J"
-        puts "  #{fg("Recent changes:", 245)}"
-        puts
-        first_change = false
-      end
-
-      timestamp = fg(Time.now.strftime("%H:%M:%S.%L"), 241)
-      display_path = fg(relative_path, 250)
-
-      case event.kind
-      when "created", "modified"
-        next unless File.exist?(file_path)
-
-        current_content = File.read(file_path)
-        previous_content = file_states[file_path]
-
-        if previous_content.nil?
-          file_states[file_path] = current_content
-          badge = bold(fg("+ added  ", 42))
-          puts "    #{timestamp} #{badge} #{display_path}"
-          next
-        end
-
-        if previous_content == current_content && !errored_files.include?(file_path)
-          next
-        end
-
-        current_parse = Herb.parse(current_content, strict: true, analyze: true)
-        current_errors = current_parse.errors
-
-        if current_errors.any?
-          previous_parse = Herb.parse(previous_content, strict: true, analyze: true)
-          previous_errors = previous_parse.errors
-
-          new_errors = current_errors.select { |error|
-            previous_errors.none? { |previous_error|
-              previous_error.error_name == error.error_name &&
-                previous_error.location.start.line == error.location.start.line
-            }
-          }
-
-          badge = bold(fg("\u{2717} error  ", 196))
-          puts "    #{timestamp} #{badge} #{display_path} #{fg("(#{current_errors.size} error#{"s" unless current_errors.size == 1})", 241)}"
-
-          new_errors.each do |error|
-            location = fg("#{relative_path}:#{error.location.start.line}:#{error.location.start.column}", 241)
-            puts "                        #{fg(error.error_name, 196)} #{location}"
-            puts "                        #{fg(error.message, 250)}" if error.message && !error.message.empty?
-          end
-
-          errored_files.add(file_path)
-
-          if websocket.client_count.positive?
-            websocket.broadcast({
-              type: "error",
-              file: relative_path,
-              errors: current_errors.map { |error|
-                {
-                  name: error.error_name,
-                  message: error.message,
-                  line: error.location.start.line,
-                  column: error.location.start.column,
-                }
-              },
-            })
-          end
-
-          puts
-          next
-        end
-
-        if errored_files.delete?(file_path)
-          file_states[file_path] = current_content
-          badge = bold(fg("\u{2713} fixed  ", 42))
-          puts "    #{timestamp} #{badge} #{display_path}"
-
-          if websocket.client_count.positive?
-            websocket.broadcast({ type: "fixed", file: relative_path })
-          end
-
-          puts
-          next
-        end
-
-        diff_result = Herb.diff(previous_content, current_content)
-        file_states[file_path] = current_content
-
-        if diff_result.identical?
-          next
-        end
-
-        operations = diff_result.operations
-
-        patchable_types = ["text_changed", "attribute_value_changed", "attribute_added", "attribute_removed"]
-
-        can_patch = operations.all? { |operation|
-          next false unless patchable_types.include?(operation.type.to_s)
-
-          next false if operation.new_node&.type&.to_s&.include?("ERB")
-          next false if operation.old_node&.type&.to_s&.include?("ERB")
-
-          true
-        }
-
-        if can_patch && websocket.client_count.positive?
-          patch_operations = operations.map do |operation|
-            {
-              type: operation.type.to_s,
-              path: operation.path,
-              old_value: extract_node_value(operation.old_node, operation.type.to_s),
-              new_value: extract_node_value(operation.new_node, operation.type.to_s),
-              old_node_type: operation.old_node&.type,
-              new_node_type: operation.new_node&.type,
-            }
-          end
-
-          websocket.broadcast({
-            type: "patch",
-            file: relative_path,
-            operations: patch_operations,
-          })
-        elsif !can_patch && websocket.client_count.positive?
-          websocket.broadcast({
-            type: "refresh",
-            file: relative_path,
-          })
-        end
-
-        badge = if can_patch
-                  bold(fg("\u{2713} patch ", 42))
-                else
-                  bold(fg("\u{21BB} refresh ", 214))
-                end
-
-        clients_label = websocket.client_count.positive? ? " #{fg("[#{websocket.client_count} client#{"s" unless websocket.client_count == 1}]", 241)}" : ""
-        puts "    #{timestamp} #{badge} #{display_path} #{fg("(#{operations.size} operation#{"s" unless operations.size == 1})", 241)}#{clients_label}"
-
-        operations.each_with_index do |operation, index|
-          type = operation.type.to_s
-          operation_path = operation.path
-          old_node = operation.old_node
-          new_node = operation.new_node
-
-          type_color = case type
-                       when "node_inserted" then 114
-                       when "node_removed", "attribute_removed" then 168
-                       when "node_replaced", "tag_name_changed" then 173
-                       when "node_wrapped", "node_unwrapped", "attribute_added", "attribute_value_changed" then 75
-                       when "node_moved" then 73
-                       when "text_changed" then 186
-                       when "erb_content_changed" then 176
-                       else 241
-                       end
-
-          type_label = type.tr("_", " ")
-          index_label = fg("##{index + 1}", 241)
-          path_label = fg("[#{operation_path.join(", ")}]", 241)
-          indent = "                        "
-
-          puts "#{indent}#{index_label} #{bold(fg(type_label, type_color))} #{path_label}"
-
-          if old_node
-            print_diff_node(indent, "-", 168, old_node, type)
-          end
-
-          next unless new_node
-
-          print_diff_node(indent, "+", 114, new_node, type)
-        end
-
-        puts
-
-      when "removed"
-        file_states.delete(file_path)
-        badge = bold(fg("- removed", 196))
-        puts "    #{timestamp} #{badge} #{display_path}"
-      end
-    end
-  rescue Interrupt
-    websocket&.stop
-    print SHOW_CURSOR
-    puts
-    puts "Stopped."
-    exit(0)
-  ensure
-    websocket&.stop
-    print SHOW_CURSOR
-  end
-
-  def print_diff_node(indent, sign, color, node, type)
-    value = extract_node_value(node, type)
-
-    if value
-      puts "#{indent}  #{fg(sign, color)} #{fg(value, color)}"
-    else
-      location = node_location_label(node)
-      puts "#{indent}  #{fg(sign, color)} #{fg(node.type, 250)}#{location}"
-    end
-  end
-
-  def node_location_label(node)
-    return "" unless node.respond_to?(:location)
-
-    " #{fg("(#{node.location.start.line}:#{node.location.start.column})", 241)}"
-  end
-
-  def extract_node_value(node, _operation_type)
-    return nil unless node
-
-    if node.is_a?(Herb::AST::HTMLTextNode)
-      return node.content&.to_s
-    end
-
-    if node.is_a?(Herb::AST::ERBContentNode)
-      return node.content&.value&.to_s
-    end
-
-    if node.is_a?(Herb::AST::HTMLElementNode)
-      name = node.tag_name
-      name = name.value if name.respond_to?(:value)
-      return "<#{name}>"
-    end
-
-    if node.is_a?(Herb::AST::HTMLAttributeNode)
-      parts = []
-
-      if node.name.respond_to?(:children)
-        parts << node.name.children.map { |child| child.respond_to?(:content) ? child.content.to_s : "" }.join
-      end
-
-      if node.value.respond_to?(:children)
-        value = node.value.children.map { |child| child.respond_to?(:content) ? child.content.to_s : "" }.join
-        parts << "=\"#{value}\""
-      end
-
-      result = parts.join
-      return result.empty? ? nil : result
-    end
-
-    nil
+    require_relative "dev/runner"
+    Herb::Dev::Runner.new(path: @file || ".").run
   end
 
   def diff_files
